# apps/web/public/journals-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenScience Journal Discovery API",
    "version": "1.0.0",
    "description": "Public journal discovery and immutable editorial interpretations. Original paper DOI identifies the scientific source; OpenScience version URLs identify derivative interpretations. This document covers the public journal extension, not private editorial operations. API requests are not AI citations."
  },
  "servers": [{"url": "/api", "description": "Same-origin web proxy"}],
  "paths": {
    "/journals": {
      "get": {
        "operationId": "listPublicJournals",
        "parameters": [
          {"$ref": "#/components/parameters/Cursor"},
          {"$ref": "#/components/parameters/Limit"},
          {"name": "query", "in": "query", "schema": {"type": "string", "maxLength": 200}},
          {"name": "subject", "in": "query", "schema": {"type": "string", "maxLength": 100}}
        ],
        "responses": {"200": {"description": "Published homepages ordered by journal UUID ascending.", "content": {"application/json": {"schema": {"type": "object", "required": ["items", "nextCursor"], "properties": {"items": {"type": "array", "items": {"$ref": "#/components/schemas/Journal"}}, "nextCursor": {"type": ["string", "null"]}}}}}}, "429": {"$ref": "#/components/responses/RateLimited"}}
      }
    },
    "/journals/{slugOrId}": {
      "get": {
        "operationId": "getPublicJournal",
        "parameters": [{"name": "slugOrId", "in": "path", "required": true, "schema": {"type": "string"}}],
        "responses": {"200": {"description": "Public journal identity.", "content": {"application/json": {"schema": {"type": "object", "properties": {"journal": {"$ref": "#/components/schemas/Journal"}}}}}}, "404": {"$ref": "#/components/responses/NotFound"}, "429": {"$ref": "#/components/responses/RateLimited"}}
      }
    },
...
  },
  "components": {
    "parameters": {
      "Cursor": {"name": "cursor", "in": "query", "description": "Use nextCursor from the preceding page.", "schema": {"type": "string", "format": "uuid"}},
      "Limit": {"name": "limit", "in": "query", "schema": {"type": "integer", "minimum": 1, "maximum": 100, "default": 20}},
...
    },
    "responses": {
      "NotFound": {"description": "Unknown or unpublished identity. Private records are not disclosed."},
      "RateLimited": {"description": "Platform rate limit reached. Respect the response Retry-After header when provided."}
    },
    "schemas": {
      "Journal": {"type": "object", "properties": {"id": {"type": "string", "format": "uuid"}, "slug": {"type": "string"}, "nameZh": {"type": ["string", "null"]}, "nameEn": {"type": ["string", "null"]}, "websiteUrl": {"type": "string", "format": "uri"}, "pIssn": {"type": ["string", "null"]}, "eIssn": {"type": ["string", "null"]}, "subjects": {"type": "array", "items": {"type": "string"}}, "status": {"enum": ["active", "paused", "closed", "reverification"]}, "publicArticleCount": {"type": "integer", "minimum": 0}}},
...
    }
  }
}